Improve try_parse error
If `get_magnitude_multiplier` cannot find a matching entry in `suffixes`, it returns `0.0`.
This is then used as a failure case in `try_parse` to specify a mismatched suffix.

This is not clear to endusers what the specific mismatch was, and what the valid cases are.

Create a new method `try_get_magnitude_multiplier` that returns a more explicit error in this case.

In addition, fix mispelling of the `get_magnitude_multiplier` method.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>